diabetsCare-poo-main/requirements.txt
numpy>=1.26
matplotlib>=3.8
Pillow>=10.0.0
